+++ crates/common/src/models/execution_log.rs
@@ -7,7 +7,7 @@ pub struct ExecutionLog {
     pub log_id: String,
     pub crdb_region: String,
     pub execution_id: String,
-    pub attempt_number: i32,
+    pub attempt_number: i64,
     pub level: String,
     pub message: String,
     pub logged_at: DateTime<Utc>,
